variable changed to double

+++ src/reader/getRowsNnzPerProc.cu
@@ -2,13 +2,13 @@
 
 void getRowsNnzPerProc(int *rowsPerGpu, int *nnzPGPU, const int *global_n, const int *global_nnz,  const int *row_Ptr, const int ngpus) 
 {
-    float nnzIncre = (float ) *global_nnz/ (float) ngpus;
-    float lookingFor=nnzIncre;
+    double nnzIncre = (double) *global_nnz/ (double) ngpus;
+    double lookingFor=nnzIncre;
     int startRow=0, endRow;
     int partition=0;
 
     for (int row=0; row<*global_n; ++row) {    
-        if ( (float) row_Ptr[row+1] >=  lookingFor ) { 
+        if ( (double) row_Ptr[row+1] >=  lookingFor ) { 
             // search for smallest difference
             if (fabs ( lookingFor - row_Ptr[row+1])  <= fabs ( lookingFor - row_Ptr[row])   ) {
                 endRow = row;
